fonction supprimer livre

+++ controllers/Stuff.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book')
+const fs = require('fs')
 
 exports.createBook = (req, res, next) => {
     const bookObject = JSON.parse(req.body.book)
@@ -55,12 +56,24 @@ exports.getOneBook = (req, res, next) => {
 }
 
 exports.modifyBook = (req, res, next) => {
-    Book.updateOne(
-        { _id: req.params.id }, 
-        { ...req.body, _id: req.params.id }
-    )
-    .then(() => res.status(201).json({ message: 'Livre modifié !' }))
-    .catch(error => res.status(400).json({ error }));
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body }
+    delete bookObject._userId
+    Book.findOne({_id: req.params.id})
+    .then((book) => {
+        if (book.userId != req.auth.userId) {
+            res.status(401).json({ message: 'Non authorisé' })
+        } else {
+            Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Livre modifié' }))
+            .catch(error => res.status(401).json({ error }))
+        }
+    })
+    .catch((error) => {
+        res.status(400).json({ error })
+    })
 }
 
 exports.rateBook = (req, res, next) => {
@@ -100,11 +113,22 @@ exports.rateBook = (req, res, next) => {
 }
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({_id: req.params.id})
-    .then(() => {
-        res.status(200).json({ message: 'Deleted !' })
+    Book.findOne({ _id: req.params.id })
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Non autorisé' })
+            } else {
+                const filename = book.imageUrl.split('/images/')[1]
+                fs.unlink(`images/${filename}`, () => {
+                     Book.deleteOne({_id: req.params.id})
+                        .then(() => {
+                            res.status(200).json({ message: 'Livre supprimé !' })})
+                        .catch((error) => {
+                            res.status(401).json({ error: error })})
+                })
+            }
     })
-    .catch((error) => {
-        res.status(400).json({ error: error })
-    })
+        .catch(error => {
+            res.status(500).json({ error })
+        })
 }
